+++ webrtc/src/webrtc_c/api/rtp_sender_interface.cc
@@ -1,6 +1,7 @@
 #include "rtp_sender_interface.h"
 
 #include <assert.h>
+#include <memory>
 
 #include <api/rtc_error.h>
 #include <api/rtp_parameters.h>
@@ -23,24 +24,22 @@ webrtc_RtpSenderInterface_GetParameters(
   return reinterpret_cast<struct webrtc_RtpParameters*>(parameters);
 }
 
-WEBRTC_EXPORT struct webrtc_RTCError_unique*
-webrtc_RtpSenderInterface_SetParameters(
+WEBRTC_EXPORT void webrtc_RtpSenderInterface_SetParameters(
     struct webrtc_RtpSenderInterface* self,
-    const struct webrtc_RtpParameters* parameters) {
-  auto sender = reinterpret_cast<webrtc::RtpSenderInterface*>(self);
+    const struct webrtc_RtpParameters* parameters,
+    struct webrtc_RTCError_unique** out_rtc_error) {
+  assert(out_rtc_error != nullptr);
   assert(parameters != nullptr);
-  if (parameters == nullptr) {
-    return reinterpret_cast<struct webrtc_RTCError_unique*>(
-        new webrtc::RTCError(webrtc::RTCErrorType::INVALID_PARAMETER,
-                             "parameters is null"));
-  }
+  auto sender = reinterpret_cast<webrtc::RtpSenderInterface*>(self);
   auto p = reinterpret_cast<const webrtc::RtpParameters*>(parameters);
   auto result = sender->SetParameters(*p);
   if (result.ok()) {
-    return nullptr;
+    *out_rtc_error = nullptr;
+  } else {
+    auto error = std::make_unique<webrtc::RTCError>(result);
+    *out_rtc_error =
+        reinterpret_cast<struct webrtc_RTCError_unique*>(error.release());
   }
-  return reinterpret_cast<struct webrtc_RTCError_unique*>(
-      new webrtc::RTCError(result));
 }
 
 WEBRTC_EXPORT int webrtc_RtpSenderInterface_SetTrack(