+++ webrtc/src/webrtc_c/api/rtp_sender_interface.h
@@ -16,10 +16,10 @@ WEBRTC_DECLARE_REFCOUNTED(webrtc_RtpSenderInterface);
 
 WEBRTC_EXPORT struct webrtc_RtpParameters*
 webrtc_RtpSenderInterface_GetParameters(struct webrtc_RtpSenderInterface* self);
-WEBRTC_EXPORT struct webrtc_RTCError_unique*
-webrtc_RtpSenderInterface_SetParameters(
+WEBRTC_EXPORT void webrtc_RtpSenderInterface_SetParameters(
     struct webrtc_RtpSenderInterface* self,
-    const struct webrtc_RtpParameters* parameters);
+    const struct webrtc_RtpParameters* parameters,
+    struct webrtc_RTCError_unique** out_rtc_error);
 WEBRTC_EXPORT int webrtc_RtpSenderInterface_SetTrack(
     struct webrtc_RtpSenderInterface* self,
     struct webrtc_MediaStreamTrackInterface* track);